implement suggestions and fix bug
